Add ":curwin" modifier to Vim
Add a new command modifier that changes how splits work, similar to
`:vertical` and `:tab`. The difference is that it will make the command
use an *existing* window rather than splitting off a new window.

The user can use [count] to specify which window to use (defaults to
current window) and [!] to forcefully discard a window with modified
buffer.

Most of the handling happens in `win_split()` where it detects the
command mod and pretends to split off a window. Commands like
`:terminal` handles the modifier manually because it already has to deal
with the `++curwin` option.

Examples of commands affected by this:

* help
* split / new
* sbuffer
* snext
* ptag
* cw / copen
* cc / cnext
* terminal

Note that there's currently a bug / issue where the "<mods>" passed into
custom commands don't retain [count] and [!] for modifiers, so you
cannot do something like `:2curwin! MyCommand` as "<mods>" will just
show "curwin". This is something that should be fixed. It also affects
`:tab`.